new architecture (in english)
