+++ .gitignore
@@ -3,3 +3,4 @@ A telecharger.txt
 *.mp3
 *.torrent
 *.txt
+*.pyc
